fix: 회사명 자동 완성을 search API 로 이전
- companies API 에 담긴 자동 완성 기능을 search API 로 이전했습니다.
- search 디렉토리와 컨트롤러, 서비스를 만들고 거기에 search_company 메소드를 작성했습니다.

resolves: #4
