+++ app/companies/companies_controller.py
@@ -1,23 +1,8 @@
-from flask import jsonify, make_response, json
+from flask import make_response
 from flask_restx import Namespace, Resource, reqparse
 from .companies_service import CompaniesService as service
 
 ns = Namespace("companies")
-
-
-@ns.route("/search")
-class SearchByQuery(Resource):
-    def get(self):
-        parser = reqparse.RequestParser()
-        parser.add_argument("query", type=str, location="args")
-        parser.add_argument("x-wanted-language", type=str, location="headers")
-        args = parser.parse_args()
-
-        result = service.search_company(args)
-        if result["ok"]:
-            return result["data"]
-        else:
-            return make_response(result["error"], result["http_status"])
 
 
 @ns.route("/<string:keyword>")
